modify: LadderGameConsole -> LadderGameConsoleController로 이름 변경
